Make clef settings uniform for all nodes

+++ cmd/beekeeper/cmd/cluster.go
@@ -35,7 +35,10 @@ func (c *command) deleteCluster(ctx context.Context, clusterName string, cfg *co
 			// delete nodes from the node group
 			g := cluster.NodeGroup(ng)
 			for i := 0; i < len(v.Nodes); i++ {
-				nName := v.Nodes[i].Name
+				nName := fmt.Sprintf("%s-%d", ng, i)
+				if len(v.Nodes[i].Name) > 0 {
+					nName = v.Nodes[i].Name
+				}
 				if err := g.DeleteNode(ctx, nName); err != nil {
 					return fmt.Errorf("deleting node %s from the node group %s", nName, ng)
 				}
@@ -53,19 +56,39 @@ func (c *command) deleteCluster(ctx context.Context, clusterName string, cfg *co
 
 			// delete nodes from the node group
 			g := cluster.NodeGroup(ng)
-			for i := 0; i < v.Count; i++ {
-				nName := fmt.Sprintf("%s-%d", ng, i)
-				if err := g.DeleteNode(ctx, nName); err != nil {
-					return fmt.Errorf("deleting node %s from the node group %s", nName, ng)
-				}
+			if len(v.Nodes) > 0 {
+				for i := 0; i < len(v.Nodes); i++ {
+					nName := fmt.Sprintf("%s-%d", ng, i)
+					if len(v.Nodes[i].Name) > 0 {
+						nName = v.Nodes[i].Name
+					}
+					if err := g.DeleteNode(ctx, nName); err != nil {
+						return fmt.Errorf("deleting node %s from the node group %s", nName, ng)
+					}
 
-				if deleteStorage && *ngConfig.PersistenceEnabled {
-					pvcName := fmt.Sprintf("data-%s-0", nName)
-					if err := c.k8sClient.PVC.Delete(ctx, pvcName, clusterOptions.Namespace); err != nil {
-						return fmt.Errorf("deleting pvc %s: %w", pvcName, err)
+					if deleteStorage && *ngConfig.PersistenceEnabled {
+						pvcName := fmt.Sprintf("data-%s-0", nName)
+						if err := c.k8sClient.PVC.Delete(ctx, pvcName, clusterOptions.Namespace); err != nil {
+							return fmt.Errorf("deleting pvc %s: %w", pvcName, err)
+						}
+					}
+				}
+			} else {
+				for i := 0; i < v.Count; i++ {
+					nName := fmt.Sprintf("%s-%d", ng, i)
+					if err := g.DeleteNode(ctx, nName); err != nil {
+						return fmt.Errorf("deleting node %s from the node group %s", nName, ng)
+					}
+
+					if deleteStorage && *ngConfig.PersistenceEnabled {
+						pvcName := fmt.Sprintf("data-%s-0", nName)
+						if err := c.k8sClient.PVC.Delete(ctx, pvcName, clusterOptions.Namespace); err != nil {
+							return fmt.Errorf("deleting pvc %s: %w", pvcName, err)
+						}
 					}
 				}
 			}
+
 		}
 	}
 
@@ -91,6 +114,7 @@ func (c *command) setupCluster(ctx context.Context, clusterName string, cfg *con
 			if !ok {
 				return nil, fmt.Errorf("node group profile %s not defined", v.Config)
 			}
+
 			if v.Mode == "bootnode" { // TODO: implement standalone mode
 				// add node group to the cluster
 				cluster.AddNodeGroup(ng, ngConfig.Export())
@@ -99,25 +123,38 @@ func (c *command) setupCluster(ctx context.Context, clusterName string, cfg *con
 				g := cluster.NodeGroup(ng)
 				errGroup := new(errgroup.Group)
 				for i := 0; i < len(v.Nodes); i++ {
-					nName := v.Nodes[i].Name
+					// set node name
+					nName := fmt.Sprintf("%s-%d", ng, i)
+					if len(v.Nodes[i].Name) > 0 {
+						nName = v.Nodes[i].Name
+					}
+					// set bootnodes
 					beeConfig, ok := cfg.BeeConfigs[v.BeeConfig]
 					if !ok {
 						return nil, fmt.Errorf("bee profile %s not defined", v.BeeConfig)
 					}
 					bConfig := beeConfig.Export()
-
 					bConfig.Bootnodes = fmt.Sprintf(v.Nodes[i].Bootnodes, clusterConfig.GetNamespace()) // TODO: improve bootnode management, support more than 2 bootnodes
 					bootnodes += bConfig.Bootnodes + " "
-					bOptions := bee.NodeOptions{
-						Config:       &bConfig,
-						ClefKey:      v.Nodes[i].ClefKey,
-						ClefPassword: v.Nodes[i].ClefPassword,
-						LibP2PKey:    v.Nodes[i].LibP2PKey,
-						SwarmKey:     v.Nodes[i].SwarmKey,
+					// set NodeOptions
+					nOptions := bee.NodeOptions{
+						Config: &bConfig,
+					}
+					if len(v.Nodes[i].Clef.Key) > 0 {
+						nOptions.ClefKey = v.Nodes[i].Clef.Key
+					}
+					if len(v.Nodes[i].Clef.Password) > 0 {
+						nOptions.ClefPassword = v.Nodes[i].Clef.Password
+					}
+					if len(v.Nodes[i].LibP2PKey) > 0 {
+						nOptions.LibP2PKey = v.Nodes[i].LibP2PKey
+					}
+					if len(v.Nodes[i].SwarmKey) > 0 {
+						nOptions.SwarmKey = v.Nodes[i].SwarmKey
 					}
 
 					errGroup.Go(func() error {
-						return g.SetupNode(ctx, nName, bOptions, fund)
+						return g.SetupNode(ctx, nName, nOptions, fund)
 					})
 				}
 
@@ -132,27 +169,59 @@ func (c *command) setupCluster(ctx context.Context, clusterName string, cfg *con
 			if !ok {
 				return nil, fmt.Errorf("node group profile %s not defined", v.Config)
 			}
+
 			if v.Mode != "bootnode" { // TODO: support standalone nodes
-				// add node group to the cluster
-				gOptions := ngConfig.Export()
-				nProfile, ok := cfg.BeeConfigs[v.BeeConfig]
+				// set bootnodes
+				beeConfig, ok := cfg.BeeConfigs[v.BeeConfig]
 				if !ok {
 					return nil, fmt.Errorf("bee profile %s not defined", v.BeeConfig)
 				}
-				nConfig := nProfile.Export()
-				nConfig.Bootnodes = bootnodes
-				gOptions.BeeConfig = &nConfig
-				cluster.AddNodeGroup(ng, gOptions)
+				bConfig := beeConfig.Export()
+				bConfig.Bootnodes = bootnodes
+				// add node group to the cluster
+				ngOptions := ngConfig.Export()
+				ngOptions.BeeConfig = &bConfig
+				cluster.AddNodeGroup(ng, ngOptions)
 
 				// start nodes in the node group
 				g := cluster.NodeGroup(ng)
 				errGroup := new(errgroup.Group)
-				for i := 0; i < v.Count; i++ {
-					nName := fmt.Sprintf("%s-%d", ng, i)
 
-					errGroup.Go(func() error {
-						return g.SetupNode(ctx, nName, bee.NodeOptions{}, fund)
-					})
+				if len(v.Nodes) > 0 {
+					for i := 0; i < len(v.Nodes); i++ {
+						// set node name
+						nName := fmt.Sprintf("%s-%d", ng, i)
+						if len(v.Nodes[i].Name) > 0 {
+							nName = v.Nodes[i].Name
+						}
+						// set NodeOptions
+						nOptions := bee.NodeOptions{}
+						if len(v.Nodes[i].Clef.Key) > 0 {
+							nOptions.ClefKey = v.Nodes[i].Clef.Key
+						}
+						if len(v.Nodes[i].Clef.Password) > 0 {
+							nOptions.ClefPassword = v.Nodes[i].Clef.Password
+						}
+						if len(v.Nodes[i].LibP2PKey) > 0 {
+							nOptions.LibP2PKey = v.Nodes[i].LibP2PKey
+						}
+						if len(v.Nodes[i].SwarmKey) > 0 {
+							nOptions.SwarmKey = v.Nodes[i].SwarmKey
+						}
+
+						errGroup.Go(func() error {
+							return g.SetupNode(ctx, nName, nOptions, fund)
+						})
+					}
+				} else {
+					for i := 0; i < v.Count; i++ {
+						// set node name
+						nName := fmt.Sprintf("%s-%d", ng, i)
+
+						errGroup.Go(func() error {
+							return g.SetupNode(ctx, nName, bee.NodeOptions{}, fund)
+						})
+					}
 				}
 
 				if err := errGroup.Wait(); err != nil {
@@ -167,6 +236,7 @@ func (c *command) setupCluster(ctx context.Context, clusterName string, cfg *con
 			if !ok {
 				return nil, fmt.Errorf("node group profile %s not defined", v.Config)
 			}
+
 			if v.Mode == "bootnode" { // TODO: support standalone nodes
 				// add node group to the cluster
 				cluster.AddNodeGroup(ng, ngConfig.Export())
@@ -174,24 +244,37 @@ func (c *command) setupCluster(ctx context.Context, clusterName string, cfg *con
 				// add nodes to the node group
 				g := cluster.NodeGroup(ng)
 				for i := 0; i < len(v.Nodes); i++ {
-					nName := v.Nodes[i].Name
+					// set node name
+					nName := fmt.Sprintf("%s-%d", ng, i)
+					if len(v.Nodes[i].Name) > 0 {
+						nName = v.Nodes[i].Name
+					}
+					// set bootnodes
 					beeConfig, ok := cfg.BeeConfigs[v.BeeConfig]
 					if !ok {
 						return nil, fmt.Errorf("bee profile %s not defined", v.BeeConfig)
 					}
 					bConfig := beeConfig.Export()
-
 					bConfig.Bootnodes = fmt.Sprintf(v.Nodes[i].Bootnodes, clusterConfig.GetNamespace()) // TODO: improve bootnode management, support more than 2 bootnodes
 					bootnodes += bConfig.Bootnodes + " "
-					bOptions := bee.NodeOptions{
-						Config:       &bConfig,
-						ClefKey:      v.Nodes[i].ClefKey,
-						ClefPassword: v.Nodes[i].ClefPassword,
-						LibP2PKey:    v.Nodes[i].LibP2PKey,
-						SwarmKey:     v.Nodes[i].SwarmKey,
+					// set NodeOptions
+					nOptions := bee.NodeOptions{
+						Config: &bConfig,
+					}
+					if len(v.Nodes[i].Clef.Key) > 0 {
+						nOptions.ClefKey = v.Nodes[i].Clef.Key
+					}
+					if len(v.Nodes[i].Clef.Password) > 0 {
+						nOptions.ClefPassword = v.Nodes[i].Clef.Password
+					}
+					if len(v.Nodes[i].LibP2PKey) > 0 {
+						nOptions.LibP2PKey = v.Nodes[i].LibP2PKey
+					}
+					if len(v.Nodes[i].SwarmKey) > 0 {
+						nOptions.SwarmKey = v.Nodes[i].SwarmKey
 					}
 
-					if err := g.AddNode(nName, bOptions); err != nil {
+					if err := g.AddNode(nName, nOptions); err != nil {
 						return nil, fmt.Errorf("adding node %s: %w", nName, err)
 					}
 				}
@@ -203,25 +286,56 @@ func (c *command) setupCluster(ctx context.Context, clusterName string, cfg *con
 			if !ok {
 				return nil, fmt.Errorf("node group profile %s not defined", v.Config)
 			}
+
 			if v.Mode != "bootnode" { // TODO: support standalone nodes
-				// add node group to the cluster
-				gOptions := ngConfig.Export()
-				nProfile, ok := cfg.BeeConfigs[v.BeeConfig]
+				// set bootnodes
+				beeConfig, ok := cfg.BeeConfigs[v.BeeConfig]
 				if !ok {
 					return nil, fmt.Errorf("bee profile %s not defined", v.BeeConfig)
 				}
-				nConfig := nProfile.Export()
-				nConfig.Bootnodes = bootnodes
-				gOptions.BeeConfig = &nConfig
+				bConfig := beeConfig.Export()
+				bConfig.Bootnodes = bootnodes
+				// add node group to the cluster
+				gOptions := ngConfig.Export()
+				gOptions.BeeConfig = &bConfig
 				cluster.AddNodeGroup(ng, gOptions)
 
 				// add nodes to the node group
 				g := cluster.NodeGroup(ng)
-				for i := 0; i < v.Count; i++ {
-					nName := fmt.Sprintf("%s-%d", ng, i)
 
-					if err := g.AddNode(nName, bee.NodeOptions{}); err != nil {
-						return nil, fmt.Errorf("adding node %s: %w", nName, err)
+				if len(v.Nodes) > 0 {
+					for i := 0; i < len(v.Nodes); i++ {
+						// set node name
+						nName := fmt.Sprintf("%s-%d", ng, i)
+						if len(v.Nodes[i].Name) > 0 {
+							nName = v.Nodes[i].Name
+						}
+						// set NodeOptions
+						nOptions := bee.NodeOptions{}
+						if len(v.Nodes[i].Clef.Key) > 0 {
+							nOptions.ClefKey = v.Nodes[i].Clef.Key
+						}
+						if len(v.Nodes[i].Clef.Password) > 0 {
+							nOptions.ClefPassword = v.Nodes[i].Clef.Password
+						}
+						if len(v.Nodes[i].LibP2PKey) > 0 {
+							nOptions.LibP2PKey = v.Nodes[i].LibP2PKey
+						}
+						if len(v.Nodes[i].SwarmKey) > 0 {
+							nOptions.SwarmKey = v.Nodes[i].SwarmKey
+						}
+
+						if err := g.AddNode(nName, bee.NodeOptions{}); err != nil {
+							return nil, fmt.Errorf("adding node %s: %w", nName, err)
+						}
+					}
+				} else {
+					for i := 0; i < v.Count; i++ {
+						nName := fmt.Sprintf("%s-%d", ng, i)
+
+						if err := g.AddNode(nName, bee.NodeOptions{}); err != nil {
+							return nil, fmt.Errorf("adding node %s: %w", nName, err)
+						}
 					}
 				}
 			}
